+++ engine/src/strategy_engine/node/indicator_node/mod.rs
@@ -33,7 +33,7 @@ pub struct IndicatorNode {
 
 impl IndicatorNode {
     pub fn new(
-        strategy_id: i32, 
+        strategy_id: i64, 
         node_id: String, 
         node_name: String, 
         exchange: Exchange,
