+++ engine/src/strategy_engine/strategy/add_if_else_node.rs
@@ -13,7 +13,7 @@ impl Strategy {
     pub async fn add_if_else_node(
         graph: &mut Graph<Box<dyn NodeTrait>, (), Directed>, 
         node_indices: &mut HashMap<String, NodeIndex>, 
-        strategy_id: i32,
+        strategy_id: i64,
         node_id: String, 
         node_name: String,
         cases: Vec<Case>,
